Rename DictionaryReqeusts model to Dictionary.
