+++ src/lib/models/api/Dictionary.ts
